[ash] Ensure nested DisplayManager changes are ordered correctly
Currently it is possible for DisplayManager observers to trigger
further synchronous DisplayManager changes during change event
propagation. This nested behavior can result in change
notifications being delivered to observers out of order.

A concrete example of where this is currently happening is in
b/327282654. This occurs as follows.

  1. A Display is added and DisplayManager begins notifying
     observers of the added Display
  2. One of the observers updates the insets of the added Display
  3. DisplayManager now notifies observers of the inset change
  4. DisplayManager notifies the remaining observers of the
     original added Display

As a result the later observers receive DisplayManager changes
out of order and invariants fail.

This CL addresses this issue for DisplayManagerObservers. A
BeginEndNotifier is created when propagating notifications so
that these can be correctly sequenced after all observers have
been notified of the original change.

The change preserves the "will process" and "did process" observer
sequencing for a given change. This is not ideal given this is
propagated after the DisplayManager has actually updated its state,
howevever preserving this sequencing seems better than delivering
multiple "will process" events followed by multiple "did process"
events.

A more correct and robust fix for this behavior will be explored in
the resolution of b/328134509.

Bug: 327282654, 328134509
Change-Id: I4d4795b2989316461fe94026e2eff5d5ee27ed86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5341339
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271084}