Scope repairing shortcuts to installers running interactively
When the installer installs the same version as the current one, we
check if the installer is running in the interactive user context. If it
is, then it is safe to assume recreating the shortcuts will be fine. If
it is not (running as system as part of a silent update), then we do not
re-create the shortcuts.

Bug: 328843997
Change-Id: Id1e8d3f0a18fa5bab1611441fe2d8de030a979dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5357741
Commit-Queue: Javier Flores Assad <floresa@microsoft.com>
Reviewed-by: Mohamed Mansour <mmansour@microsoft.com>
Reviewed-by: S Ganesh <ganesh@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271110}