trusted_vault: report more information about recovery factors.
When Chrome receives an indication that it should recheck the
hw_protected security domain it'll need to know whether a) the security
domain has been reset and b) whether the GPM PIN has been changed.

(In the future it might also need to know whether there's an iCloud
Keychain recovery factor in the domain too.)

Because of all that, this change makes the result of
DownloadAuthenticationFactorsRegistrationState a struct and expands it
with the members that we'll want in the short term.

Change-Id: I67eeb0810429fc9bc5209ff662f0b1b881b87cd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5353695
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271131}