Use scoped_refptr<Adapters> in SessionImpl.
This allows the Adaptor to be used for the full life of the session,
even if adapters are loaded from disk.

Previously, if the adapters were reloaded, the adapter could be missing
from the interpreter (perhaps briefly) or may not be the one the same as
the one the session was started with.

Change-Id: I25dd6d194b1e4db7551c6c061f1984ec577ac78f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5349933
Commit-Queue: Steven Holte <holte@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271145}