[Rounded Windows] Use window radius property to get the radius of window
Currently, we use the helper method chromeos::GetWindowCornerRadius()
to get the radius of window.

In this CL, we migrate to use the window property,
`aura::client::kWindowCornerRadiusKey` to get the radius of the window.
Note: ShadowController already uses this property to rounded the window
drop shadow.

The benefit of this change is that, we can easily control rounding of
window and its decoration thus allowing us to control rounding of
different types of Windows. This behavior is important for exo-based
windows where the client applications can specify their rounding.

Bug: b/327961488
Change-Id: Ie337663b01673866fe4c85e37ec40bd2161b4f82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5338146
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271149}