Remove donnd@chromium.org from chrome/android/java/src/org/chromium/chrome/browser/contextualsearch/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/donnd@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iadf5f568aaba71831989968278f4e767be826988
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5359976
Reviewed-by: Gang Wu <gangwu@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271160}