Re-introduce span::operator==
This allows spanification to change `memcmp(ptr, ptr, size) == 0` to
become simply `span == span`.

The operator differs from std::ranges::equal in a few ways:
- It's much simpler to type/find.
- It fails comparison of fixed-size spans that have different sizes
  at compile time (similar to in Rust) instead of returning false
  at runtime.

The operator differs from memcmp() in a few ways:
- It's not a three-way comparison, which can be added later.
- It allows comparison between spans of different types, as long as
  the types are comparable (they satisfy the concept
  std::equality_comparable_with).

This operator was previously removed in 03fe4779b619e78b since the
operator was removed from std::span, in order to ease transition from
base::span to std::span, however that is no longer our plan, as we are
extending span in many ways to make it work better for spanification
at scale.

Unlike the previous implementation of the operator, it now supports
implicit conversions to span. This allows comparing spans with other
things that can become a span, such as arrays, including implicit
const promotion in the conversion.

The removal of the operator== and operator<=> greatly increased the
verbosity of code that had been using them before.

Notably absl::Span still has the operator, and this will be useful
for making spanification easier for codebases that are built on
absl, and this change helps make things at least as easy for Chrome.

While much has been said on the regularity of types, I just straight
up disagree that view types should suffer from API challenges based
on this, and that their regularity should be invalidated based on the
fact that C++ allows aliased mutation. View types should be treated
as regular under the assumption that the backing store they point to
is not invalidated while they are in use. Academics aside, this will
greatly simplify how spanification changes are written that involve
comparison.

Bug: 40284755
Change-Id: Ida57c6e890bafa26e30eb0b3ca2ad286d7b8790f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5353552
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271188}