Add arg to resolve DUT hostnames in ChromeOS Swarming.
When running tasks on ChromeOS Swarming, there's no special hostname that resolves to the current DUT's hostname. As a workaround, we can parse the hostname out of SWARMING_BOT_ID env var. In this CL, we define an arg to tell test_runner to parse the hostname.

Bug: b/313960680
Change-Id: Ie822f75673b020fd9d0ecb55f3803332660162e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5346288
Commit-Queue: Eduardo Yap <eduardoyap@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271241}