[commerce] Allow product info to fall back to on-demand if possible
This patch allows the GetProductInfoForUrl API in the shopping service
to fall back to the on-demand API that optimization guide provides.
This will only be called if the user is known to have visited the URL
that is being queried. In our case, this currently corresponds to open
tabs but will also reference sync entities in the future.

Bug: b:328622410
Change-Id: Icecb8abb971f1860c00477e4a70d05cc26f0aa3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5356564
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Ayman Almadhoun <ayman@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271254}