[Panorama] Hide overflow-x
A horizontal scrollbar is showing at minimum width.

There seems to be an extra width on #container. The extra width goes
away with the removal of border-box, but then we get double vertical
scrollbars. I removed the overflow: auto on #container to handle it
instead. It will use the styling of .sp-scroller instead, which sets
overflow-x to hidden.

Also, removed GM2 margin-bottom (it was being overridden by the GM3
style below it so it went unnoticed) that was accidentally left in.

Bug: 5319610
Change-Id: Id43666421a93ead51e452110b2f5e10dd609419b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362799
Commit-Queue: Riley Tatum <rtatum@google.com>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271274}