[Benefits][Settings] Clear benefits when benefits toggle is turned off
This CL addresses the card benefit toggle state change event. When the
toggle is switched off, it updates the kAutofillPaymentCardBenefits
preference to false. The PaymentsDataManager monitors changes in the
kAutofillPaymentCardBenefits preference state. Upon detecting a change,
it triggers the OnAutofillPaymentsCardBenefitsPrefChange method, which
clears all benefits if the preference is set to false.

This CL refactors the benefit clearing process, moving it from the
PersonalDataManager to the PaymentsDataManager.
  Adds PaymentsDataManagerTestApi.
  Removes the ClearAllCreditCardBenefits method from the PDM.

Design: go/pay-autofill-chrome-benefits-design

Bug: 324094766
Change-Id: I782a242d06c288d6f69545dd8e8ec3cdc7232d91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5330170
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Viplav Kadam <viplavkadam@google.com>
Reviewed-by: Olivia Saul <jsaul@google.com>
Reviewed-by: Alexander Tekle <alexandertekle@google.com>
Cr-Commit-Position: refs/heads/main@{#1271284}