Remove angelaxiao@chromium.org from tools/metrics/histograms/metadata/quick_answers/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/angelaxiao@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ife93781022ca618a5152ea309dea4669f63c4597
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5352475
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271341}