Record delay between FCP frame received to surface embedding in blink
The first contentful paint frame for a new document might be submitted
before or after the surface used for it is embedded by the browser
process. This CL makes us record the delay in both cases, to figure
out how often each case happens and how long the delay is between
the events.

Bug: 325798142
Change-Id: I698b6b43d81b6ec7a764e283ceb34dae6e19ca5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318409
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271343}