[SWAutoPreload] Make IOBuffer size configurable
This CL makes buffer size for ReadData configurable. As the provided
DataPipeAllocationSize::kLargerSizeIfPossible is too large for android
devices, this CL introduces `read_buffer_size` to decide the buffer
size.

Also, this CL introduce the boolean param, `query_data_size`, which gets
the upcoming data size in advance, and creates IOBuffer with the actual
size. This approach is more reasonable rather than having the magic
number, but it increase the number of DataPipe calls, so implemented
behind the feature param.

Bug: 329014514
Change-Id: I1f8a2c9e84d560ab782b5cbce78acb4ba07b6128
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5359598
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271356}