[FspCache] Introduce map of operation IDs to file handles
Previously, the caller of a CloudFileSystem::OpenFile request would be
returned the file handle created by the underlying FSP OpenFile request.

Instead, the CloudFileSystem should assign a new operation ID upon a
successful underlying FSP OpenFile request. This operation ID will be
returned to the caller and will be used by the caller to communicate
with the CloudFileSystem regarding the opened file.

The CloudFileSystem will still use the file handle to communicate with
the underlying FSP by mapping each incoming operation ID to the file
handle created by the underlying FSP. This mapping will be deleted upon
file close.

The motivation for this change is so that in the future, a unique
mapping can be made between disk backed operations (identified by a disk
FD) and cloud backed FSP operations (identified by a file handle). The 2
operations may return identifiers that clash and any requests for a
clashing identifier would be difficult to disambiguate.

Test: unit_tests --gtest_filter=FileSystemProviderCloudFileSystemTest*
Bug: b/328670254
Change-Id: I9884f90ebbbcc6b2e4d1aa2628a79560c98858eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5354173
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Reviewed-by: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271374}