[Files]: Define the index interface and uses a delegate.
Introduces FileIndex interface. Creates an in memory implementation of
it. Converts FileIndexService to just a facade that relies on the
delegate to implement them. Introduces a method that allows us to add
more terms without removing them. Introduces namespaces, which map to
the field name.

Bug: b:327535200
Change-Id: I68d88469844aa98cb443134910532b180c06eb82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5359357
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Bo Majewski <majewski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271386}