[FspCache] Add CacheManagerObserver
Add a CacheManager observer called CacheManagerObserver. This observes
when providers are initialised through the
CacheManager::InitializeForProvider() method.

This allows FileSystemProviderCloudFileSystemTest unit tests to be able
to wait on the asynchronous initialisation of the CloudFileSystem before
proceeding with the test. The asynchronous part is the
CacheManager::InitializeForProvider() call. This functionality will be
used in the tests of http://crrev.com/c/5362655.

Bug: b/328679875
Change-Id: I626e4e0a8768bea2dbce5d4f361e29b48841748d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362654
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Cr-Commit-Position: refs/heads/main@{#1271407}