Printing: Render document contents before headers / footers.
This is what the spec expects. It currently shouldn't affect the output,
though, since the document contents, which is rendered by Blink,
currently cannot overlap with the headers and footers. However, doing it
in this order will be necessary once we add support for @page
backgrounds, which should be painted under headers and footers, not on
top of them.

This change affects one accessibility test, though, which assumed that
the text that it was looking for was the last accessibility node, which
will no longer be the case.

Bug: 40286153
Change-Id: Ic751bcf5ff0ab331d66846b082e72d5fbb9a4959
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5360895
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271420}