[iOS] Don't use all items in the TabStrip layout
This CL rolls back a change that, when scrolling, returned the
attributes for all the elements in the collection view and not just
the visible ones.
Rolled back CL: crrev.com/c/5234136

This would be too costly in terms of performances, so it changes it
back.

This is creating a small issue when scrolling fast a very large number
of tabs where the last separator is not shown.

Bug: 41481434
Change-Id: Iebac3e8fa628696ccae7d92e27dab82db7dc440c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5365081
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1271464}