[ChromeSigninBubble] Update the intercept logic
Added a new pref to remember the user choice on the bubble if
accept/decline is chosen. Also added a new pref to count the number of
dimisses. Those two prefs are stored per account/profile.

Based on the user choice, the next web sign in will be impacted:
- accepted: lead to automatic sign in.
- declined: leads to no sign in and no show of the bubble.
- dismissed: count the number of dismisses. When the max is reached,
treat the result as a decline. Max is currently set to 5.

Also changed the way Uno responds to ACCESS_POINT_WEB_SIGNIN:
Since we now remember the user choice, a sign in with an
ACCESS_POINT_WEB_SIGNIN could result in an automatic signin
based on the user previous choice from settings/bubble.
Also we do not reset the explicit browser signin pref, instead we
just keep it as is.

Added browser tests to cover the new behavior.

Also split test into Full and/or experimental modes for clarity:
- Tests with full and experimental mode: some tests are needed for
both modes with some slight differences.
- Tests with full mode only, changes that do not affect the experiment.
- Tests with experiment mode only, tests the old behavior that got updated in this CL (That should be removed when we stop supporting
the experiment soon).

Added `Signin.Intercept.ChromeSignin.DismissesBeforeAccept/Decline` which replaced `Signin.Intercept.ChromeSignin.AttemptsBeforeAccept` with a more accurate count given the new behavior. Keeping the old
histogram for the experiment phase that is still on going.

Bug: b:319396084
Change-Id: I7611ae6db30e08940de4973f8c21aa66c3da7980
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5326931
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1271465}