[iOS] Update TabStrip move
This CL updates how the TabStrip is handling the move operation.
In particular, it is adding a method to move the item instead of
reloading all of them.

It is also fixing a bug where the check to know if a cell was reused
or not was using the visibility of the cell in the Collection View.
This is not working for drag and drop as the cell is not considered
visible when dropped, so there was an early return creating an empty
favicon.

It is also fixing a few stylistic issues wrt to first letter being
lowercase.

Fixed: 326240736, 326241280, 326241080
Change-Id: I38c0e194cc0d703315d2a232b557af0ba9622e9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5355701
Reviewed-by: Quentin Pubert <qpubert@google.com>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271485}