[//media] Enable 2c pure-SW VideoFrame->GL upload w/Metal, non-OOP-C
This CL enables 2-copy upload of pure software VideoFrames to GL when
ANGLE-Metal and non-OOP-C are being used. The thing that we need to do
in this case is create the intermediate SharedImage that holds the YUV
data of the VideoFrame via GLTextureImageBacking rather than
IOSurfaceImageBacking, as the latter doesn't support I420 as a format.

GLTextureImageBackingFactory rejects creation of
SharedImages with GLES2 usage when ANGLE-Metal is being used, as those
could be used on a different GPU than raster. However, in this case
the GLES2 usage is present *only* because we know that raster is going
over the GLES2 interface - the intermediate SI is not actually accessed
by WebGL, as it is first copied to an RGBA SI that itself is then copied
to a GL texture via WebGL's GLES2 interface. GLTextureImageBacking will
allow creation of SharedImages if it knows that the SI is in this case
via the client specifying SHARED_IMAGE_USAGE_GLES2_FOR_RASTER_ONLY.

For Graphite, we have an additional issue wherein we disallow
GLTextureImageBacking creation for SIs with raster usage, since raster
might not be using GL. However, in this case we *know* that raster is
going over the GLES2 interface, and thus it is fine to use
GLTextureImageBacking. GLTexture similarly will allow creation of
SharedImages in this case via the client specifying the
RASTER_OVER_GLES2_ONLY flag.

This CL adds both usages to the intermediate SI, as well as unittests
of GLTextureImageBackingFactory that specifically test that it allows
creation of SharedImages matching the parameters given for that
intermediate SI.

I have verified that this CL fixes the Google Earth crash from [1] with
Ganesh and Graphite.

I will also be developing a regression test of the high-level use case
here (import of pure-SW video frames to WebGL).

[1] https://issues.chromium.org/u/1/issues/328302269

Bug: 328472684
Change-Id: Ia66460f99bc770170a2fef681620808acf975d35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5349642
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271500}