Skip processing DispatchLoad calls on commit-intercepted RFHs
In tests, we can intercept DidCommit navigation callbacks to keep the
navigation in a "pending commit" stage. If this happens on a new RFH,
the RFH will also stay in the kPendingCommit state. However, since
the renderer thought it already sent the DidCommit IPC and the RFH
is now active, we might receive some IPCs that should only be called
on active RFHs, e.g. DispatchLoad, causing failed DCHECKs.

In this CL, we keep track of whether we've intercepted a commit on
a RFH before, and skip processing DispatchLoad in that case. This
shouldn't have any effect outside of tests.

Bug: 936696
Change-Id: If2f9ce61532f9213c04bb5343e945de4ebb9735b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5364257
Reviewed-by: Fergal Daly <fergal@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271504}