[//gpu] Destroy known WebGPU devices in WebGPUDecoderImpl destruction
We want to have WebGPUDecoderImpl destroy any created WebGPU devices
when it is destroyed. This will allow device destruction to propagate
throughout the system. Concrete motivation is that an upcoming change is
to going to have IOSurfaceImageBacking hold a ref to the device(s) via
SharedTextureMemory. With that change we will have IOSurfaceImageBacking
check when it should drop SharedTextureMemory instances due to the
corresponding device being lost. However, if WebGPUDecoderImpl doesn't
destroy its created devices when it goes away, then those devices will
never *be* lost as long as the IOSurfaceImageBacking is alive. This
could be problematic in the case where IOSurfaceImageBacking(s) are used
in a pool.

Bug: 1515822
Change-Id: Idfb7bb7006ba295f07cadd79e70a18b5684e0052
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5364262
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271576}